fuck PJ64 1.6, rip 64 bit timer

+++ src/game/game_init.c
@@ -67,7 +67,7 @@ u8 gIsConsole;
 u8 gWidescreen;
 #endif
 u8 gCustomCameraMode;
-u64 gSpeedrunTimer;
+u32 gSpeedrunTimer;
 u16 sCurrFBNum = 0;
 u16 frameBufferIndex = 0;
